+++ src/main/java/net/silentchaos512/gear/block/analyzer/TilePartAnalyzer.java
@@ -18,16 +18,25 @@
 
 package net.silentchaos512.gear.block.analyzer;
 
+import com.google.common.collect.ImmutableList;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
 import net.minecraft.network.NetworkManager;
 import net.minecraft.network.play.server.SPacketUpdateTileEntity;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.tags.Tag;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ITickable;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.util.LazyOptional;
+import net.minecraftforge.items.CapabilityItemHandler;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.wrapper.SidedInvWrapper;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.parts.IGearPart;
 import net.silentchaos512.gear.api.parts.MaterialGrade;
@@ -38,16 +47,24 @@ import net.silentchaos512.lib.tile.SyncVariable;
 import net.silentchaos512.lib.tile.TileSidedInventorySL;
 import net.silentchaos512.lib.util.TimeUtils;
 
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.List;
 import java.util.stream.IntStream;
 
 public class TilePartAnalyzer extends TileSidedInventorySL implements ITickable {
     static final int BASE_ANALYZE_TIME = TimeUtils.ticksFromSeconds(5);
+    private static final List<Tag<Item>> CATALYST_TAGS = ImmutableList.of(
+            new ItemTags.Wrapper(SilentGear.getId("analyzer_catalyst/tier1")),
+            new ItemTags.Wrapper(SilentGear.getId("analyzer_catalyst/tier2"))
+    );
 
     static final int INPUT_SLOT = 0;
-    private static final int[] SLOTS_INPUT = {INPUT_SLOT};
-    private static final int[] SLOTS_OUTPUT = IntStream.rangeClosed(1, 4).toArray();
+    static final int CATALYST_SLOT = 1;
+    private static final int[] SLOTS_INPUT = {INPUT_SLOT, CATALYST_SLOT};
+    private static final int[] SLOTS_OUTPUT = {2, 3, 4, 5};
     static final int INVENTORY_SIZE = SLOTS_INPUT.length + SLOTS_OUTPUT.length;
+    private static final int[] SLOTS_ALL = IntStream.rangeClosed(0, INVENTORY_SIZE).toArray();
 
     @SyncVariable(name = "progress")
     int progress = 0;
@@ -82,8 +99,15 @@ public class TilePartAnalyzer extends TileSidedInventorySL implements ITickable 
                     stack.setCount(1);
                     input.shrink(1);
 
+                    // Get catalyst
+                    ItemStack catalyst = getCatalystStack();
+                    int catalystTier = getCatalystTier(catalyst);
+                    if (catalystTier > 0 && !catalyst.isEmpty()) {
+                        catalyst.shrink(1);
+                    }
+
                     // Assign grade, move to output slot
-                    MaterialGrade.selectRandom(SilentGear.random).setGradeOnStack(stack);
+                    MaterialGrade.selectWithCatalyst(SilentGear.random, catalystTier).setGradeOnStack(stack);
                     setInventorySlotContents(outputSlot, stack);
                     if (input.getCount() <= 0) {
                         for (int slot : SLOTS_INPUT) {
@@ -118,13 +142,15 @@ public class TilePartAnalyzer extends TileSidedInventorySL implements ITickable 
     }
 
     private ItemStack getInputStack() {
-        for (int slot : SLOTS_INPUT) {
-            ItemStack stack = getStackInSlot(slot);
-            if (PartManager.from(stack) instanceof PartMain && MaterialGrade.fromStack(stack) == MaterialGrade.NONE) {
-                return stack;
-            }
+        ItemStack stack = getStackInSlot(INPUT_SLOT);
+        if (PartManager.from(stack) instanceof PartMain && MaterialGrade.fromStack(stack) == MaterialGrade.NONE) {
+            return stack;
         }
         return ItemStack.EMPTY;
+    }
+
+    private ItemStack getCatalystStack() {
+        return getStackInSlot(CATALYST_SLOT);
     }
 
     private int getFreeOutputSlot() {
@@ -136,28 +162,20 @@ public class TilePartAnalyzer extends TileSidedInventorySL implements ITickable 
         return -1;
     }
 
-    @Override
-    public int getField(int id) {
-        switch (id) {
-            case 0:
-                return progress;
-            default:
-                return 0;
-        }
+    static boolean isUngradedMainPart(ItemStack stack) {
+        MaterialGrade grade = MaterialGrade.fromStack(stack);
+        if (grade != MaterialGrade.NONE) return false;
+        return PartManager.from(stack) instanceof PartMain;
     }
 
-    @Override
-    public void setField(int id, int value) {
-        switch (id) {
-            case 0:
-                this.progress = value;
-                break;
+    static int getCatalystTier(ItemStack stack) {
+        Item item = stack.getItem();
+        for (int i = 0; i < CATALYST_TAGS.size(); ++i) {
+            if (item.isIn(CATALYST_TAGS.get(i))) {
+                return i + 1;
+            }
         }
-    }
-
-    @Override
-    public int getFieldCount() {
-        return 1;
+        return 0;
     }
 
     @Override
@@ -206,21 +224,31 @@ public class TilePartAnalyzer extends TileSidedInventorySL implements ITickable 
     @Override
     public int[] getSlotsForFace(EnumFacing side) {
         switch (side) {
+            case UP:
+                return SLOTS_INPUT.clone();
             case DOWN:
                 return SLOTS_OUTPUT.clone();
             default:
-                return SLOTS_INPUT.clone();
+                return SLOTS_ALL.clone();
         }
     }
 
     @Override
     public boolean isItemValidForSlot(int index, ItemStack stack) {
-        if (index != INPUT_SLOT || stack.isEmpty() || (!getStackInSlot(index).isEmpty() && !getStackInSlot(index).isItemEqual(stack)))
+        if (index != INPUT_SLOT && index != CATALYST_SLOT) {
             return false;
+        }
 
-        MaterialGrade grade = MaterialGrade.fromStack(stack);
-        if (grade != MaterialGrade.NONE) return false;
-        return PartManager.from(stack) instanceof PartMain;
+        ItemStack stackInSlot = getStackInSlot(index);
+        if (stack.isEmpty() || (!stackInSlot.isEmpty() && !stackInSlot.isItemEqual(stack))) {
+            return false;
+        }
+
+        if (index == INPUT_SLOT) {
+            return isUngradedMainPart(stack);
+        } else {
+            return getCatalystTier(stack) > 0;
+        }
     }
 
     @Override
@@ -231,6 +259,22 @@ public class TilePartAnalyzer extends TileSidedInventorySL implements ITickable 
     @Override
     public boolean canExtractItem(int index, ItemStack stack, EnumFacing direction) {
         return index != INPUT_SLOT;
+    }
+
+    private final LazyOptional<? extends IItemHandler>[] handlers =
+            SidedInvWrapper.create(this, EnumFacing.UP, EnumFacing.DOWN, EnumFacing.NORTH);
+
+    @Nonnull
+    @Override
+    public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable EnumFacing side) {
+        if (!this.removed && side != null && cap == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY) {
+            if (side == EnumFacing.UP)
+                return handlers[0].cast();
+            if (side == EnumFacing.DOWN)
+                return handlers[1].cast();
+            return handlers[2].cast();
+        }
+        return super.getCapability(cap, side);
     }
 
     @Override